requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-allauth==0.51.0
django-environ==0.9.0
django-extensions==3.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
Markdown==3.4.1
oauthlib==3.2.0
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.4.2
tzdata==2022.1
django-allauth==0.51.0
drf_social_oauth2==1.2.1
drf-yasg==1.21.3
urllib3==1.26.11

